Hotfix T164 (Babel not active)
Summary:
Issue was a race condition between 2 threads. `waitUntil` for important variables being set should solve it.
Not exactly a fancy fix, but since task setup etc. might do weird stuff, I didn't want to mess with the architecture of the whole thing.

Test Plan:
Uhm..
It's a rare and hard to reproduce issue..
Test with one person on dedicated succeeded, so it's probably not game breaking.
Should try to use it in a proper mission though.

Reviewers: jsm, Walker

Reviewed By: jsm

Maniphest Tasks: T164

Differential Revision: https://dev.beowulfso.com/D5